version: Move package to internal.
